requirements.txt
langchain
langchain_community
langchain_groq
langchain_huggingface
streamlit
sentence-transformers
python-dotenv
